Fix some stuttering animations

+++ src/ui_connection_widget.hh
@@ -43,7 +43,7 @@ struct ConnectionWidget : Widget {
     float radar_alpha_target = 0;
     float prototype_alpha = 0;
     float prototype_alpha_target = 0;
-    float time_seconds = 0;
+    double time_seconds = 0;
   };
 
   static ConnectionWidget* Find(Location& here, Argument& arg);
